feat(localization): removed localization for user notification in new group
