+++ app/src/main/kotlin/org/coremapx/app/localization/objects/DialogsKeys.kt
@@ -1,0 +1,8 @@
+package org.coremapx.app.localization.objects
+
+object DialogsKeys {
+    // User Notification
+    const val USER_NOTIFICATION_SAVE_ERROR = "userNotificationSaveError"
+    const val USER_NOTIFICATION_SAVE_SUCCESS = "userNotificationSaveSuccess"
+    const val USER_NOTIFICATION_SAVE_SUCCESS_MESSAGE = "userNotificationSaveSuccessMessage"
+}
