+++ app/src/main/kotlin/org/coremapx/app/localization/objects/LocalizationKeys.kt
@@ -2,4 +2,5 @@ package org.coremapx.app.localization.objects
 
 object LocalizationKeys {
     val ui = UIKeys
+    val dialogs = DialogsKeys
 }
