+++ app/src/main/kotlin/org/coremapx/app/localization/states/LocalizationStates.kt
@@ -2,4 +2,5 @@ package org.coremapx.app.localization.states
 
 class LocalizationStates {
     val ui = UIState()
+    val dialogs = DialogsState()
 }
